@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=77.0.0", "setuptools_scm[toml]>=6.4.0"]
+requires = ["setuptools>=80.9.0", "setuptools_scm[toml]>=6.4.0"]
 backend-path = ["dissect/util"]
 build-backend = "_build"
 
@@ -7,7 +7,7 @@ build-backend = "_build"
 name = "dissect.util"
 description = "A Dissect module implementing various utility functions for the other Dissect modules"
 readme = "README.md"
-requires-python = "~=3.9"
+requires-python = ">=3.10.0"
 license = "Apache-2.0"
 license-files = ["LICENSE", "COPYRIGHT"]
 authors = [
@@ -32,12 +32,33 @@ homepage = "https://dissect.tools"
 documentation = "https://docs.dissect.tools/en/latest/projects/dissect.util"
 repository = "https://github.com/fox-it/dissect.util"
 
+
+[dependency-groups]
+test = [
+    "pytest",
+]
+lint = [
+    "ruff==0.12.12",
+    "vermin",
+]
+build = [
+    "build",
+]
+debug = [
+    "ipdb",
+]
+dev = [
+    {include-group = "test"},
+    {include-group = "lint"},
+    {include-group = "debug"},
+]
+
 [project.scripts]
 dump-nskeyedarchiver = "dissect.util.tools.dump_nskeyedarchiver:main"
 
 [tool.ruff]
 line-length = 120
-required-version = ">=0.11.0"
+required-version = ">=0.12.0"
 
 [tool.ruff.format]
 docstring-code-format = true
@@ -83,7 +104,7 @@ ignore = ["E203", "B904", "UP024", "ANN002", "ANN003", "ANN204", "ANN401", "S110
 
 [tool.ruff.lint.per-file-ignores]
 "tests/_docs/**" = ["INP001"]
-"tests/**" = ["S101"]
+"tests/**" = ["S101", "PLC0415"]
 
 [tool.ruff.lint.isort]
 known-first-party = ["dissect.util"]
