@@ -2,7 +2,7 @@ use pyo3::prelude::*;
 
 mod compression;
 
-#[pymodule]
+#[pymodule(gil_used = false)]
 fn _native(m: &Bound<'_, PyModule>) -> PyResult<()> {
     compression::create_submodule(m)?;
     Ok(())
